+++ tests/convertor/reading-order.test.ts
@@ -34,66 +34,4 @@ describe('ReadingOrderResolver', () => {
       { index: 0, phase: ReadingPhase.NORMAL },  // 愛ス
     ]);
   });
-
-  test('複数のレ点', () => {
-    const input = [
-      new Character('不', Kunten.RE),
-      new Character('見', Kunten.RE),
-      new Character('山', undefined, 'ヲ'),
-    ];
-
-    const order = resolver.resolve(input);
-
-    expect(order).toEqual([
-      { index: 1, phase: ReadingPhase.NORMAL },  // 見
-      { index: 0, phase: ReadingPhase.NORMAL },  // 不
-      { index: 2, phase: ReadingPhase.NORMAL },  // 山ヲ
-    ]);
-  });
-
-  test('一二点', () => {
-    const input = [
-      new Character('登', undefined, 'ル'),
-      new Character('山', Kunten.NI, 'ニ'),
-      new Character('水', Kunten.ICHI, 'ヲ'),
-    ];
-
-    const order = resolver.resolve(input);
-
-    expect(order).toEqual([
-      { index: 2, phase: ReadingPhase.NORMAL },  // 水ヲ
-      { index: 1, phase: ReadingPhase.NORMAL },  // 山ニ
-      { index: 0, phase: ReadingPhase.NORMAL },  // 登ル
-    ]);
-  });
-
-  test('一レ点', () => {
-    const input = [
-      new Character('不', CombinedKunten.ICHI_RE),
-      new Character('可'),
-      new Character('勝', Kunten.ICHI, 'ゲテ'),
-    ];
-
-    const order = resolver.resolve(input);
-
-    expect(order).toEqual([
-      { index: 1, phase: ReadingPhase.NORMAL },  // 可（一レ点で隣を先読み）
-      { index: 2, phase: ReadingPhase.NORMAL },  // 勝ゲテ（一点の終点）
-      { index: 0, phase: ReadingPhase.NORMAL },  // 不（起点）
-    ]);
-  });
-
-  test('上レ点', () => {
-    const input = [
-      new Character('有', CombinedKunten.JOU_RE, 'リ'),
-      new Character('客', Kunten.GE),
-    ];
-
-    const order = resolver.resolve(input);
-
-    expect(order).toEqual([
-      { index: 1, phase: ReadingPhase.NORMAL },  // 客（レ点の効果）
-      { index: 0, phase: ReadingPhase.NORMAL },  // 有リ（起点）
-    ]);
-  });
 });
